Refactor recursive _align_segment as iterative
Recursion is neat, but Python is not the best language for it. There
is no real need to use it here, so we break the recursive call out
into an interative loop instead. Only the `start` of the remainder
changes, so we increment it same as the recursive call, and once we
get to a segment that no longer needs to be further split, we
`return` (raising `StopIteration`), ending the generator.

The corresponding test was written to make the recursive approach
fall over after the default recursion depth of 1000 was hit. In
that implementation, the number of sub-segments was equal to the
eventual recursion depth. The test as-written will generate 65536
sub-segments, well beyond the recursive capabilities, but handled
fine by the iterative approach. If we make the segment upper bound
much larger, we start slowing down the tests, so this value seems
like a good compromise.