merge: support 'identical' on overlapped data
When merging multiple .hex files, in case of overlap we sometimes need to
just 'ensure' that values are identical.

The new 'identical' mode ensures this.

Note: the behavior for start_addr is currently a 'identical' (i.e. the
overlap configuration is checked only when values differ). So we might
want to change the behavior for data too:
- if data is identical: silently bail out
- add a 'stricter' mode that corresponds to the current behavior