Use entrypoints for the scripts
Swap out `scripts` for entrypoints to ensure executable flag is properly
set when installing from wheel, and hashbang is correct for whatever
virtualenv/pipenv/etc the user might be user (instead of hardcoded to
`/usr/bin/python` and relying on `script` rewrite magic to fix it).

Fix a deprecation notice for setup.cfg.

_Note: moving `scripts` to `intelhex/scripts` means that the scripts
cannot be run as-is from their position in the file system. Recommended
approach when developing a package is to `pip install -e .` install in
editable mode and work from there._