Make __getitem__ max address aware
In order to be able to convert IntelHex objects to lists, make
__getitem__ aware of the current maximum address. Yield IndexErrors for
breaching the maximum to make list generation finite. Return padding
bytes for unset addresses below the maximum, as before.

Note that this change now requires an address to be explicitly set after
object initialization to return anything but an IndexError.

Fixes #54